feat(controller): renomeia ProfileController para ProfilesController e adiciona lógica de acesso ao perfil fix(routes): altera a rota de profile para perfil e ajusta o controlador correspondente feat(view): adiciona a visualização do perfil do usuário na nova estrutura de controller test: adiciona testes para a nova página de perfil e remove testes obsoletos
