+++ app/controllers/application_controller.rb
@@ -2,7 +2,7 @@ class ApplicationController < ActionController::Base
   include Authentication
   # Only allow modern browsers supporting webp images, web push, badges, import maps, CSS nesting, and CSS :has.
   allow_browser versions: :modern
-  helper_method :current_user
+  helper_method :current_user, :current_profile
 
   private
 
@@ -12,5 +12,9 @@ class ApplicationController < ActionController::Base
 
   def current_session
     @current_session ||= Session.find_by(id: cookies.signed[:session_id])
+  end
+
+  def current_profile
+    @profile ||= current_user&.profile
   end
 end
