+++ app/views/layouts/_navbar.html.erb
@@ -21,11 +21,7 @@
             <a href="#"><%= current_user&.profile&.fullname.presence || 'Sem nome' %></a>
             <div class="uk-navbar-dropdown">
               <ul class="uk-nav uk-navbar-dropdown-nav uk-text-white">
-                <% if current_user.profile %>
-                  <li>
-                    <%= link_to 'Meu perfil', profile_path(current_user.profile.slug) %>
-                  </li>
-                <% end %>
+                <%= link_to 'Meu perfil', profile_path %>
                 <li class="uk-active"><a href="/settings/edit/settings.html">Configurações</a></li>
                 <li>
                   <%= link_to 'Sair', session_path, data: { turbo_method: :delete } %>
