+++ spec/routes/profile_spec.rb
@@ -5,10 +5,9 @@ require "rails_helper"
 RSpec.describe "Content", type: :routing do
   context "GET /profile/john-doe" do
     it do
-      expect(get: "/profile/john-doe").to route_to(
-        controller: "profile",
-        action: "show",
-        slug: "john-doe",
+      expect(get: "/perfil").to route_to(
+        controller: "profiles",
+        action: "show"
       )
     end
   end
